isis: extract graph builder into graph.rs
Move LspMap, ReachMap, ReachMapV6, graph(), graph_mt2() and their
private vertex/link helpers out of inst.rs into a new graph.rs.
inst.rs drops from 3295 to 2835 lines; the new module is 474 lines
that own everything LSDB-to-SPF-graph related.

Public surface stays the same — siblings (inst.rs, link.rs, tilfa.rs)
now import from super::graph instead of super::inst.
