+++ zebra-rs/src/isis/mod.rs
@@ -47,3 +47,5 @@ pub mod flood;
 pub use flood::LspFlood;
 
 pub mod tilfa;
+
+pub mod graph;
